Update the example App
so it reflects the changes in the Readme.md
