+++ example-app/AndroidManifest.xml
@@ -6,8 +6,8 @@
 
     <uses-permission android:name="android.permission.INTERNET" /> 
 
-    <application android:label="@string/app_name" android:icon="@drawable/icon">
-        <activity android:name="ExampleApp"
+    <application android:name=".ExampleApplication" android:label="@string/app_name" android:icon="@drawable/icon">
+        <activity android:name="ExampleActivity"
                   android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
